requirements.txt
imageio==2.5.0
matplotlib==3.1.3
numpy==1.18.5
opencv-python==4.2.0.34
Pillow==8.0.1
scikit-image==0.17.2
scikit-learn==0.22.1
scipy==1.4.1
tensorflow-addons==0.11.2
tensorflow-gpu==2.3.1
